Kind of fixes #8.
This was not really an issue with libalsa or spotifyd. The problem was with
PulseAudio/ALSA interop. Since spotifyd uses libalsa, systems that use
PulseAudio need to be set up so that audio going through ALSA gets piped to
PulseAudio, according to the good people of #alsa on freenode.
There is no real fix for this, but I modified the readme to mention this.