add CI and CD workflows
This commit adds a CI workflow to the project. It checks for the correct code style (using cargo fmt) and lints the whole project as well (using cargo clippy). The last step checks if the binary can be compiled on our target OS (using cargo check).

The commit also adds a workflow for CD. After each release, the workflow will build binaries for both Linux and macOS in two flavours: slim and full. Slim does contain nothing but the proper audio backend. Full contains all features that work correctly on the specific platform. For Linux that is dbus_keyring and dbus_mpris. For macOS it is only dbus_keyring. The resulting binaries are compressed into a .tar.gz archive and uploaded as assets to the appropiate release. The script does also create sha512 checksums and uploads them as well.