fix structopt listing use-keyring as an argument instead of completely omitting it
Prior to the extra proc_macro on the struct field, use-keyring was still listed on the help dialog as a argument under the ARGS section instead of being completely omitted. The commit introduces another proc_macro that applies if the dbus_keyring feature is not used and marks the field to be skipped by serde and structopt.