add CD support for armv6 and armv7 (#479)
This PR allows us to move away from Travis and have our CD pipeline available through GitHub.

The additions include support for a `full` and `slim` version on the armv7 architecture as well as a `slim` version for the armv6 architecture. A `full` version is missing due to cross-compilation toolchain limitations.