spotifyd.service is an user service
As an user service, it can't depend on the system service `network-online.service`. Also, fail the service right away if it's not the correct path. And assume a better place for the binary (distro packages can change this anyway).

This is intended run as an user service, another .service can be proposed for a system-wide service, with sandboxing settings.