doc: brief OpenBSD bits
Users are always advised to use official binary packages and/or build
spotifyd themselves using official ports infrastructure (from which
packages are built), so haven't duplicated any compilation instructions
here.

The official package also ships a README with a suggestion on how to
configure and run `spotifyd` which ought to be enough, anyways.