fix alsa mixer calculations (#1229)
fixes #1222

I hope, this fixes the issues mentioned with `alsa` and `alsa_linear`.
Please note that one of them will behave weirdly on your system
depending on, whether your alsa control is linear or not. (At least that
is what I assume based on my experience...) However, they should both be
able to reach 0% and not be reset when hitting play.